pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["ebrahim ayyad <ayyad_ebrahim@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
